rename stepFormField and remove interfaces
