+++ protocol-designer/src/components/StepEditForm/StepEditForm.css
@@ -168,8 +168,8 @@
 }
 
 .advanced_settings_icon {
-  height: 3rem;
-  width: 3rem;
+  height: 2.25rem;
+  width: 2.25rem;
 }
 
 .section_divider {
