+++ protocol-designer/src/components/StepEditForm/fields/FlowRate/index.js
@@ -2,9 +2,7 @@
 import * as React from 'react'
 import FlowRateInput from './FlowRateInput'
 import {connect} from 'react-redux'
-import {
-  actions as steplistActions,
-} from '../../../../steplist'
+import {actions as steplistActions} from '../../../../steplist'
 import {selectors as stepFormSelectors} from '../../../../step-forms'
 import type {StepFieldName} from '../../../../steplist/fieldLevel'
 import type {BaseState, ThunkDispatch} from '../../../../types'
