+++ protocol-designer/src/components/StepEditForm/fields/WellOrder/WellOrderModal.js
@@ -16,9 +16,9 @@ import modalStyles from '../../../modals/modal.css'
 import {actions} from '../../../../steplist'
 import {selectors as stepFormSelectors} from '../../../../step-forms'
 import type {BaseState} from '../../../../types'
-import WellOrderViz from './WellOrderViz'
 import type {WellOrderOption} from '../../../../form-types'
 
+import WellOrderViz from './WellOrderViz'
 import styles from './WellOrderInput.css'
 
 const DEFAULT_FIRST: WellOrderOption = 't2b'
