+++ protocol-designer/src/components/forms/forms.css
@@ -30,3 +30,7 @@
 .row_wrapper {
   lost-utility: clearfix;
 }
+
+.dropdown_field {
+  font-size: var(--fs-body-1);
+}
