+++ protocol-designer/src/components/forms/index.js
@@ -1,5 +1,3 @@
 // @flow
 
-export {default as InputField} from './InputField'
-export {default as CheckboxField} from './CheckboxField'
-export {default as DropdownField} from './DropdownField'
+// TODO: BC 2019-1-25 put ButtonRow and other PD specific form components here
